requirements.txt
asgiref==3.5.2
Django==4.0.6
Pillow==9.2.0
sqlparse==0.4.2

pip~=21.3.1
wheel~=0.37.1
setuptools~=60.2.0
django-debug-toolbar~=3.6.0